New modules of dbse and util (#1255)
