+++ src/raspicam/RaspiCamControl.h
@@ -184,7 +184,7 @@ typedef enum
 void raspicamcontrol_check_configuration(int min_gpu_mem);
 
 int raspicamcontrol_parse_cmdline(RASPICAM_CAMERA_PARAMETERS *params, const char *arg1, const char *arg2);
-void raspicamcontrol_display_help();
+//void raspicamcontrol_display_help();
 int raspicamcontrol_cycle_test(MMAL_COMPONENT_T *camera);
 
 int raspicamcontrol_set_all_parameters(MMAL_COMPONENT_T *camera, const RASPICAM_CAMERA_PARAMETERS *params);
