+++ src/video_bktr.c
@@ -24,6 +24,8 @@
 /* For rotation */
 #include "translate.h"
 #include "motion.h"
+#include "util.h"
+#include "logger.h"
 #include "rotate.h"
 #include "video_common.h"
 #include "video_bktr.h"
