+++ src/video_v4l2.c
@@ -25,6 +25,8 @@
 */
 #include "translate.h"
 #include "motion.h"
+#include "util.h"
+#include "logger.h"
 #include "rotate.h"
 #include "video_common.h"
 #include "video_v4l2.h"
