+++ src/webu_html.c
@@ -47,6 +47,8 @@
  */
 
 #include "motion.h"
+#include "util.h"
+#include "logger.h"
 #include "webu.h"
 #include "webu_html.h"
 #include "translate.h"
