+++ src/webu_text.c
@@ -33,6 +33,8 @@
  */
 
 #include "motion.h"
+#include "util.h"
+#include "logger.h"
 #include "webu.h"
 #include "webu_text.h"
 #include "translate.h"
